Create tictactoe
Just for fun project. The code did not completed 100%. As a beginner, took about 1 week to develop the code, so its time to move on to the next lesson. 95% perfection. The code still had the original frame un-removed, which is supposed to be implemented with the working tictactoe board. Also, did not include the case if the character input by user is more than 4 character. Assuming winner can be decided from just 4 move maximum.